requirements.txt
loguru
python-dotenv
flet
